refactor: search component paths
