--- sample-qdrant.ts
@@ -1,14 +1,0 @@
-// import { QdrantClient } from "@qdrant/js-client-rest";
-
-// const client = new QdrantClient({
-// 	url: "https://009760b7-c6e9-4de5-b327-b8c198e777da.us-west-1-0.aws.cloud.qdrant.io:6333",
-// 	apiKey:
-// 		"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2Nlc3MiOiJtIn0.GkE0Lo4vj5fFnWEY17nPQTsSwkk8FBEx6PitTna4VuE",
-// });
-
-// try {
-// 	const result = await client.getCollections();
-// 	console.log("List of collections:", result.collections);
-// } catch (err) {
-// 	console.error("Could not get collections:", err);
-// }
